+++ api/weatherHook/structure.go
@@ -1,6 +1,0 @@
-package weatherHook
-
-type WeatherHook struct {
-	Location string `json:"location"`
-	Timeout  int64  `json:"timeout"`
-}