+++ ChangeLog.md
@@ -212,4 +212,8 @@
 
 [3.27.0]    **development**:    Changed Ticketmaster endpoint to be able to be called from another function, started moving structs to separete file
 
+[3.28.0]    **development**:    Moved weatherHook, changed trigger to callUrl.
 
+[3.29.0]    **development**:    Added date validation in weatherEvent.POST()
+
+[3.30.0]    **development**:    Restructured so all structs have their own file in that same package